# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PromptRequest {
  /**
     * @minLength 1
     * @maxLength 12000
     */
  input: string;
  /** @maxLength 80 */
  promptType?: string;
  /** @maxLength 80 */
  preset?: string;
  /** @maxLength 120 */
  action?: string;
}

export interface ImproveRequest {
  /**
     * @minLength 1
     * @maxLength 12000
     */
  input: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  action: string;
  /** @maxLength 80 */
  preset?: string;
}

export interface PlaygroundRequest {
  /**
     * @minLength 1
     * @maxLength 20000
     */
  prompt: string;
  /** @maxLength 12000 */
  input?: string;
  /** @maxLength 80 */
  preset?: string;
}

export interface PromptResult {
  title: string;
  prompt: string;
  promptType: string;
  intent: string;
  context: string;
  requirements: string[];
  constraints: string[];
  outputFormat: string;
  missingInformation: string[];
  suggestions: string[];
  qualityNotes: string[];
}

export interface AnalysisResult {
  intent: string;
  promptType: string;
  clarity: string;
  structure: string;
  context: string;
  ambiguity: string[];
  requirements: string[];
  constraints: string[];
  outputRequirements: string[];
  weaknesses: string[];
  suggestions: string[];
}

/**
 * @nullable
 */
export type PlaygroundResultUsage = {
  promptTokens?: number;
  completionTokens?: number;
  totalTokens?: number;
} | null;

export interface PlaygroundResult {
  response: string;
  model: string;
  generationTimeMs: number;
  /** @nullable */
  usage?: PlaygroundResultUsage;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}