# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisResult {
  intent: string;
  promptType: string;
  clarity: string;
  structure: string;
  context: string;
  ambiguity: string[];
  requirements: string[];
  constraints: string[];
  outputRequirements: string[];
  weaknesses: string[];
  suggestions: string[];
}