# lib/api-zod/src/generated/types/improveRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImproveRequest {
  /**
     * @minLength 1
     * @maxLength 12000
     */
  input: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  action: string;
  /** @maxLength 80 */
  preset?: string;
}