# lib/api-zod/src/generated/types/playgroundRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlaygroundRequest {
  /**
     * @minLength 1
     * @maxLength 20000
     */
  prompt: string;
  /** @maxLength 12000 */
  input?: string;
  /** @maxLength 80 */
  preset?: string;
}