# lib/api-zod/src/generated/types/playgroundResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaygroundResultUsage } from './playgroundResultUsage';

export interface PlaygroundResult {
  response: string;
  model: string;
  generationTimeMs: number;
  /** @nullable */
  usage?: PlaygroundResultUsage;
}