# lib/api-zod/src/generated/types/promptRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PromptRequest {
  /**
     * @minLength 1
     * @maxLength 12000
     */
  input: string;
  /** @maxLength 80 */
  promptType?: string;
  /** @maxLength 80 */
  preset?: string;
  /** @maxLength 120 */
  action?: string;
}