# lib/api-zod/src/generated/types/promptResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PromptResult {
  title: string;
  prompt: string;
  promptType: string;
  intent: string;
  context: string;
  requirements: string[];
  constraints: string[];
  outputFormat: string;
  missingInformation: string[];
  suggestions: string[];
  qualityNotes: string[];
}